Instruction struct type change

+++ include/decode.h
@@ -16,9 +16,9 @@ typedef struct
 	long value;
 } instruction_field;
 
-int handle_sign(char *param, instruction *instr);
+int16_t handle_sign(char *param, instruction *instr);
 void build_instruction_bin(instruction_field *arr, size_t field_count, uint8_t *bin_arr, char *hex_arr);
-void finalize_signed_value(int *value, instruction *instr);
+void finalize_signed_value(int16_t *value, instruction *instr);
 
 void build_branch_r1_instruction(instruction *instr, int op_code);
 void build_branch_r2_instruction(instruction *instr, int op_code);
